requirements/staging.txt
-r base.txt # includes all dependencies in `base.txt`
psycopg2-binary==2.9.11
django-debug-toolbar==6.1.0
